+++ encore_api_cli/__init__.py
@@ -1,3 +1,3 @@
 """Command Line Interface for AnyMotion API."""
 
-__version__ = "0.6.4"
+__version__ = "0.6.5"