+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Change Log
 
+## 0.6.5
+
+Released 2020-01-22
+
+- Improved output with `--verbose` option.
+- Avoided unnecessary requests to get access token.
+
 ## 0.6.4
 
 Released 2020-01-21